+++ db/MongoDatabase.js
@@ -29,6 +29,7 @@ module.exports = class MongoDatabase extends Base {
         this.settings.options = {
             bufferMaxEntries: 0,
             keepAlive: true,
+            readPreference: 'primary',
             useNewUrlParser: true,
             useUnifiedTopology: true,
             ...this.settings.options
@@ -71,82 +72,92 @@ module.exports = class MongoDatabase extends Base {
         return this._tableMap[name];
     }
 
+    startSession () {
+        this.traceCommand('Start session');
+        return this._client.startSession();
+    }
+
+    endSession (session) {
+        this.traceCommand('End session');
+        session.endSession();
+    }
+
     // OPERATIONS
 
     create (table) {
-        this.logCommand('create', {table});
+        this.traceCommand('create', {table});
         return this._connection.createCollection(table);
     }
 
-    find (table, query) {
-        this.logCommand('find', {table, query});
-        return this.getTable(table).find(query).toArray();
+    find (table, query, options) {
+        this.traceCommand('find', {table, query});
+        return this.getTable(table).find(query, options).toArray();
     }
 
     distinct (table, key, query, options) {
-        this.logCommand('distinct', {table, query});
+        this.traceCommand('distinct', {table, key, query});
         return this.getTable(table).distinct(key, query, options);
     }
 
-    async insert (table, data) {
-        this.logCommand('insert', {table, data});
+    async insert (table, data, options) {
+        this.traceCommand('insert', {table, data});
         if (Array.isArray(data)) {
-            const result = await this.getTable(table).insertMany(data);
+            const result = await this.getTable(table).insertMany(data, options);
             return result.insertedIds;
         }
-        const result = await this.getTable(table).insertOne(data);
+        const result = await this.getTable(table).insertOne(data, options);
         return result.insertedId;
     }
 
-    upsert (table, query, data) {
-        this.logCommand('upsert', {table, query, data});
-        return this.getTable(table).updateOne(query, {$set: data}, {upsert: true});
+    upsert (table, query, data, options) {
+        this.traceCommand('upsert', {table, query, data});
+        return this.getTable(table).updateOne(query, {$set: data}, {upsert: true, ...options});
     }
 
-    update (table, query, data) {
-        this.logCommand('update', {table, query, data});
-        return this.getTable(table).updateOne(query, {$set: data});
+    update (table, query, data, options) {
+        this.traceCommand('update', {table, query, data});
+        return this.getTable(table).updateOne(query, {$set: data}, options);
     }
 
-    updateAll (table, query, data) {
-        this.logCommand('updateAll', {table, query, data});
-        return this.getTable(table).updateMany(query, {$set: data});
+    updateAll (table, query, data, options) {
+        this.traceCommand('updateAll', {table, query, data});
+        return this.getTable(table).updateMany(query, {$set: data}, options);
     }
 
-    updateAllPull (table, query, data) {
-        this.logCommand('updateAllPull', {table, query, data});
-        return this.getTable(table).updateMany(query, {$pull: data});
+    updateAllPull (table, query, data, options) {
+        this.traceCommand('updateAllPull', {table, query, data});
+        return this.getTable(table).updateMany(query, {$pull: data}, options);
     }
 
-    updateAllPush (table, query, data) {
-        this.logCommand('updateAllPush', {table, query, data});
-        return this.getTable(table).updateMany(query, {$push: data});
+    updateAllPush (table, query, data, options) {
+        this.traceCommand('updateAllPush', {table, query, data});
+        return this.getTable(table).updateMany(query, {$push: data}, options);
     }
 
-    unset (table, query, data) {
-        this.logCommand('unset', {table, query, data});
-        return this.getTable(table).updateOne(query, {$unset: data});
+    unset (table, query, data, options) {
+        this.traceCommand('unset', {table, query, data});
+        return this.getTable(table).updateOne(query, {$unset: data}, options);
     }
 
-    unsetAll (table, query, data) {
-        this.logCommand('unsetAll', {table, query, data});
-        return this.getTable(table).updateMany(query, {$unset: data});
+    unsetAll (table, query, data, options) {
+        this.traceCommand('unsetAll', {table, query, data});
+        return this.getTable(table).updateMany(query, {$unset: data}, options);
     }
 
-    delete (table, query = {}) {
-        this.logCommand('delete', {table, query});
-        return this.getTable(table).deleteMany(query);
+    delete (table, query = {}, options) {
+        this.traceCommand('delete', {table, query});
+        return this.getTable(table).deleteMany(query, options);
     }
 
     async drop (table) {
         if (await this.isTableExists(table)) {
-            this.logCommand('drop', {table});
+            this.traceCommand('drop', {table});
             return this.getTable(table).drop();
         }
     }
 
     async dropAll () {
-        this.logCommand('dropAll');
+        this.traceCommand('dropAll');
         for (const name of await this.getTableNames()) {
             await this.drop(name);
         }
@@ -158,7 +169,7 @@ module.exports = class MongoDatabase extends Base {
 
     async rename (table, target) {
         if (await this.isTableExists(table)) {
-            this.logCommand('rename', {table, target});
+            this.traceCommand('rename', {table, target});
             return this._connection.renameCollection(...arguments);
         }
     }
@@ -166,7 +177,7 @@ module.exports = class MongoDatabase extends Base {
     // AGGREGATE
 
     count (table, query) {
-        this.logCommand('count', {table, query});
+        this.traceCommand('count', {table, query});
         return this.getTable(table).countDocuments(query);
     }
 
@@ -174,7 +185,7 @@ module.exports = class MongoDatabase extends Base {
 
     async queryAll (query) {
         const cmd = await this.buildQuery(query);
-        const cursor = this.getTable(cmd.from).find(cmd.where);
+        const cursor = this.getTable(cmd.from).find(cmd.where, query.getOptions());
         if (cmd.select) {
             cursor.project(cmd.select);
         }
@@ -187,7 +198,7 @@ module.exports = class MongoDatabase extends Base {
         if (cmd.limit) {
             cursor.limit(cmd.limit);
         }
-        this.logCommand('find', cmd);
+        this.traceCommand('find', cmd);
         let docs = await cursor.toArray();
         if (!cmd.order) {
             docs = query.sortOrderByKeys(docs);
@@ -213,7 +224,7 @@ module.exports = class MongoDatabase extends Base {
 
     async queryDistinct (query, key) {
         const cmd = await this.buildQuery(query);
-        return this.distinct(cmd.from, key, cmd.where, {});
+        return this.distinct(cmd.from, key, cmd.where, query.getOptions());
     }
 
     async queryScalar (query, key) {
@@ -223,32 +234,32 @@ module.exports = class MongoDatabase extends Base {
 
     async queryInsert (query, data) {
         const cmd = await this.buildQuery(query);
-        return this.insert(cmd.from, data);
+        return this.insert(cmd.from, data, query.getOptions());
     }
 
     async queryUpdate (query, data) {
         const cmd = await this.buildQuery(query);
-        return this.update(cmd.from, cmd.where, data);
+        return this.update(cmd.from, cmd.where, data, query.getOptions());
     }
 
     async queryUpdateAll (query, data) {
         const cmd = await this.buildQuery(query);
-        return this.updateAll(cmd.from, cmd.where, data);
+        return this.updateAll(cmd.from, cmd.where, data, query.getOptions());
     }
 
     async queryUpsert (query, data) {
         const cmd = await this.buildQuery(query);
-        return this.upsert(cmd.from, cmd.where, data);
+        return this.upsert(cmd.from, cmd.where, data, query.getOptions());
     }
 
     async queryDelete (query) {
         const cmd = await this.buildQuery(query);
-        return this.delete(cmd.from, cmd.where);
+        return this.delete(cmd.from, cmd.where, query.getOptions());
     }
 
     async queryCount (query) {
         const cmd = await this.buildQuery(query);
-        return this.count(cmd.from, cmd.where);
+        return this.count(cmd.from, cmd.where, query.getOptions());
     }
 
     // INDEXES
@@ -263,28 +274,42 @@ module.exports = class MongoDatabase extends Base {
      */ 
     async createIndex (table, data) {
         await this.create(table);
-        this.logCommand('createIndex', {table, data});
+        this.traceCommand('createIndex', {table, data});
         return this.getTable(table).createIndex(...data);
     }
 
     async dropIndex (table, name) {
         if (await this.isTableExists(table)) {
-            this.logCommand('dropIndex', {table, name});
+            this.traceCommand('dropIndex', {table, name});
             return this.getTable(table).dropIndex(name);
         }
     }
 
     async dropIndexes (table) {
         if (await this.isTableExists(table)) {
-            this.logCommand('dropIndexes', {table});
+            this.traceCommand('dropIndexes', {table});
             return this.getTable(table).dropIndexes();
         }
     }
 
     async reindex (table) {
         if (await this.isTableExists(table)) {
-            this.logCommand('reindex', {table});
+            this.traceCommand('reindex', {table});
             return this.getTable(table).reIndex();
+        }
+    }
+
+    async transact (handler) {
+        if (!this.enableTransactions) {
+            return handler();
+        }
+        const session = this.startSession();
+        session.startTransaction();
+        try {
+            await handler({session});
+        } finally {
+            await session.abortTransaction();
+            this.endSession(session);
         }
     }
 };